+++ NEWS.md
@@ -4,6 +4,8 @@
 * Added as.data.frame.gtypes.
 * Removed gtypes2df.
 * Added arguments to as.matrix.gtypes to include id and strata columns in output.
+* Removed the jmodeltest function as this functionality is available in the modeltest function in the phangorn package.
+* Added conversion functions gtypes2phyDat and phyDat2gtypes to facilitate interoperability with the phangorn package.
 
 ## Changes to strataG v 1.0