+++ NEWS.md
@@ -6,6 +6,7 @@
 * Added arguments to as.matrix.gtypes to include id and strata columns in output.
 * Removed the jmodeltest function as this functionality is available in the modeltest function in the phangorn package.
 * Added conversion functions gtypes2phyDat and phyDat2gtypes to facilitate interoperability with the phangorn package.
+* Removed read.arlequin.
 
 ## Changes to strataG v 1.0