+++ NEWS.md
@@ -7,6 +7,7 @@
 * Removed the jmodeltest function as this functionality is available in the modeltest function in the phangorn package.
 * Added conversion functions gtypes2phyDat and phyDat2gtypes to facilitate interoperability with the phangorn package.
 * Removed read.arlequin.
+* Added alleleNames accessor for gtypes object, which returns list of allele names for each locus.
 
 ## Changes to strataG v 1.0